fix: improve code and fix tests
There were a couple of places where the code was not as robust as it
should. In those places it was easy to get panic errors.

I rewrote the tests to remove those errors and improved a bit the
overall quality so this can be used as the foundation for future e2e
tests.

Signed-off-by: David Moreno García <damoreno@redhat.com>